+++ experimental/client-go/bindings/c/examples/train.c
