+++ experimental/client-go/bindings/core/embed-core.bin
