+++ experimental/client-go/bindings/core/generate.go
