+++ experimental/client-go/bindings/core/partialdata.go
