+++ experimental/client-go/bindings/cpp/examples/train.cpp
