+++ experimental/client-go/bindings/cpp/examples/train_session.cpp
