+++ experimental/client-go/cmd/example/.gitignore
