+++ experimental/client-go/internal/gowandb/internal_runopts/runopts.go
@@ -1,8 +1,8 @@
 package internal_runopts
 
 import (
-	"github.com/wandb/wandb/core/pkg/gowandb/opts/runopts"
 	"github.com/wandb/wandb/core/pkg/service"
+	"github.com/wandb/wandb/experimental/client-go/pkg/opts/runopts"
 )
 
 func WithTelemetry(telemetry *service.TelemetryRecord) runopts.RunOption {
