+++ experimental/client-go/pkg/opts/runopts/runopts.go
@@ -2,8 +2,8 @@
 package runopts
 
 import (
-	"github.com/wandb/wandb/core/pkg/gowandb/runconfig"
 	"github.com/wandb/wandb/core/pkg/service"
+	"github.com/wandb/wandb/experimental/client-go/pkg/runconfig"
 )
 
 type RunParams struct {
