+++ experimental/client-rust/src/printer.rs
